Improve error reporting when verifying command exec output
Using Eventually with a matcher reports that the operation timed out
which isn't very useful. Ideally, we should report the actual and
expected output to aid in debugging. To achieve this, we can actually
make assertions in the function passed into Eventually by having the
function take a single Gomega argument (see
https://onsi.github.io/gomega/#category-3-making-assertions-eminem-the-function-passed-into-codeeventuallycode).
Added an "awaitCmdOutputContains" function to encapsulate this.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>